+++ src/sdk/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Issue templates (#34) [ypcode - Yannick Plennevaux] / [jansenbe - Bert Jansen]
 - Enable PnPContext creation from an Office 365 Group id (#39) [jansenbe - Bert Jansen]
 - Added the Group model, represents an Office 365 group (#48) [jansenbe - Bert Jansen]
+- Added custom exception handling for authentication errors (#45) [jansenbe - Bert Jansen]
 
 ### Changed